tasks/parallel-web-scraping-aws/scraper/requirements.txt
boto3
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
python-dotenv
requests